requirements.txt
numpy==1.19.1
joblib==0.16.0
tqdm==4.46.0
torch==1.6.0+cu101
torchvision==0.7.0+cu101
tensorboardX==2.1
tensorboard==1.15.0
jsonpickle==1.4.1
jsonschema==3.2.0
scikit-learn==0.22.2.post1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
dgl-cu101==0.4.3.post2
graphviz==0.14.1
scipy==1.5.0
